AI: can we streamline the edit-file just like the create_file tool, in the front end ?
like stream the code-edit output of the agent; Then we should be able to show the diff as well, when the tool complete. We get the original file content, and the full updated code (output of morph) and some how send it to the front end.

the front end should show this properly , concisely so user can see the changes in green / red.

This shouldn't change the content feed to the model btw. Like it will pollute the context.

make a plan what to do first, not make changes yet